Refactor: python, cts: Move PartialStart into the _cts module.
No code changes.
